Move busy handling into LayerPass and cleanup
CURA-1601
